feat: add dialogs in Book Page

+++ MyShop/MyShop/ViewModel/BooksViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using CommunityToolkit.Mvvm.Input;
 using MyShop.Model;
 using MyShop.Repository;
+using MyShop.Services;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -38,26 +39,35 @@ namespace MyShop.ViewModel
 
         public void ExecuteEditBookCommand()
         {
+            if (SelectedBook == null)
+            {
+                App.MainRoot.ShowDialog("No selected item", "Please select an item first!");
+                return;
+            }
             ParentPageNavigation.ViewModel = new EditBookViewModel(SelectedBook);
+
+            App.MainRoot.ShowDialog("Success", "Book is updated!");
         }
 
         public async void ExecuteDeleteBookCommand()
         {
-            //Message to user to delete or not
+            if (SelectedBook == null)
             {
-
+                await App.MainRoot.ShowDialog("No selected item", "Please select an item first!");
+                return;
             }
+            var confirmed = await App.MainRoot.ShowYesCancelDialog("Delete this item?", "Delete", "Cancel");
 
-            //Delete
+            if (confirmed == true)
             {
                 var task = await _bookRepository.Remove(SelectedBook.Id);
                 if (task)
                 {
-                    //Success message
+                    await App.MainRoot.ShowDialog("Success", "Book is removed!");
                 }
                 else
                 {
-                    //Failed message
+                    await App.MainRoot.ShowDialog("Failure", "Removal unsuccessful...");
                 }
             }
             
@@ -66,6 +76,8 @@ namespace MyShop.ViewModel
         public void ExecuteAddBookCommand()
         {
             ParentPageNavigation.ViewModel = new AddBookViewModel();
+
+            App.MainRoot.ShowDialog("Success", "Book is added!");
         }
         
         public async void ExecuteGetAllCommand()
@@ -73,8 +85,6 @@ namespace MyShop.ViewModel
             var task = await _bookRepository.GetAll();
             Books = task;
         }
-
-
 
     }
 }
